refactor and some comments
