+++ pkg/window/strategy/sliding/sliding.go
@@ -25,9 +25,9 @@ import (
 	"sync"
 	"time"
 
-	slist "github.com/numaproj/numaflow/pkg/shared/slist"
 	"github.com/numaproj/numaflow/pkg/window"
 	"github.com/numaproj/numaflow/pkg/window/keyed"
+	"github.com/numaproj/numaflow/pkg/window/wlist"
 )
 
 // Sliding implements sliding windows
@@ -47,7 +47,7 @@ type Sliding struct {
 	// the traversal from the tail of the list for Get and Create Operations. For Remove Operations, since
 	// the earlier windows are expected to be closed before the more recent ones, we start the traversal
 	// from the Head.
-	entries *slist.SortedWindowList[window.AlignedKeyedWindower]
+	entries *wlist.SortedWindowList[window.AlignedKeyedWindower]
 	lock    sync.RWMutex
 }
 
@@ -58,7 +58,7 @@ func NewSliding(length time.Duration, slide time.Duration) *Sliding {
 	return &Sliding{
 		Length:  length,
 		Slide:   slide,
-		entries: slist.New[window.AlignedKeyedWindower](),
+		entries: wlist.New[window.AlignedKeyedWindower](),
 		lock:    sync.RWMutex{},
 	}
 }
